Added support for finding the intersection of line segments.
Created the package deltagruppen.circles.math and added the  LineSegment and LineSegmentPair classes to it. These classes are to be used when trying to find an imperfect circle in a list of points.

I also added the class LineSegmentPairTest which is a JUnit test class, and changed the app config to enable the running of unit tests. Unit testing in AS is experimental so you might have to jump through some extra hoops to run the tests. [This](http://tools.android.com/tech-docs/unit-testing-support) link has a step-by-step guide to setting it up.