Imperfect circles can now be recognised.
The ImperfectCircle class represents a user-drawn imperfect circle. It takes a list of points, turns that into a list of line segments and tries to find two that intersect. If it succeeds the user did in fact draw a circle, if not it throws an error.

The DrawView class has been edited to try to create an imperfect circle on the touch up event, and print the success/failure of the attempt to logcat. So now we can distinguish circles from user input, yay! :)