Fix for issue #19.
Some devices (including our test device) register the same point twice in a row on touch move events, which messes with our algorithms for detecting the imperfect circle because we get line segments of zero length.

The `cleanInput()` method in this fix goes through the list of points and removes such duplicates.

Link to the issue: https://github.com/deltagruppen/circles/issues/19