Beautified the Pi Info activity.
Completely reworked the layout of the Pi Info activity. Moved to a strictly two-column layout with the title in the first column. Put the fact list in tables so that the numbering is formatted correctly (i.e. no text under the numbers). Added a splash of color for good measure.

Also, the facts are now split up into different <string> elements in strings.xml. This let me be more flexible with the formatting.